refactor(core): rehome decision_router upward-import surfaces (VAL-P4A-003)
Move the two surfaces aragora.core.decision_router imported from aragora.server
down to their canonical lower-layer homes so core no longer reaches up into the
server layer at any scope:
- aragora/server/documents.py -> aragora/documents/parsing.py
- aragora/server/decision_integrity_utils.py -> aragora/pipeline/decision_integrity_utils.py

Both old paths keep importing via re-export shims that emit DeprecationWarning,
so the ~15 existing consumers are unaffected. decision_router now imports the
parse helpers from aragora.documents.parsing and build_decision_integrity_payload
from aragora.pipeline.decision_integrity_utils. Repoint the decision-integrity
unit test's patch targets + the path-traversal test's private-helper imports to
the canonical modules.
